+++ .env
@@ -4,6 +4,4 @@ POSTGRES_PASSWORD="masterkey"
 POSTGRES_HOST="postgres"
 POSTGRES_PORT='5432'
 POSTGRES_DSN='postgres://postgres:masterkey@localhost:5432/test_aiohttp'
-REBBITMQ_URL='http://localhost:5672'
-RABBITMQ_DEFAULT_USER="payments"
-RABBITMQ_DEFAULT_PASS="hackme"
+RABBITMQ_DSN ="amqp://payments:hackme@localhost:5672"
