+++ Readme.md
@@ -1,0 +1,12 @@
+- следи за названиями reseive(поправил receive) и это помоему не единственное место
+- папка app_rabbitMQ наталкивает на мысль что должен быть app а когда заходишь одни worker которые не понятно что делают
+- put/receive это по своей сути один и тот же функционал его можно в один класс плюс гораздо наглядней работа с коннекшином
+- вынес в базовый конфиг твой в settings тут хоть видно какие данные нужны для старта проекта
+- у меня чет проблема не могу pull сделать образа я его поменял на который был скачен
+- и согласись async with RabbitClient() as connection: выглядит компактнее чем было 
+- еще у тебя должно подчеркивать что классы названы camelcase обрати внимание как назвал
+- для автоформатирования строк используй command + shift + L будет по pep8 все хотябы
+- не забывай про гитигнор 
+
+ну вот так будем потхоньку работать
+давай теперь потренируемся и сделаем merge в ветку main
