+++ docker-compose.yaml
@@ -1,7 +1,7 @@
 version: "3.2"
 services:
   rabbitmq:
-    image: rabbitmq:3-management-alpine
+    image: rabbitmq:3.8-management
     container_name: 'rabbitmq'
     ports:
         - 5672:5672
@@ -9,6 +9,8 @@ services:
     environment:
       - RABBITMQ_DEFAULT_USER=payments
       - RABBITMQ_DEFAULT_PASS=hackme
+      - RABBITMQ_DEFAULT_VHOST=default_host
+
 
 
 #  backend:
