+++ requierments.txt
@@ -12,3 +12,5 @@ multidict==6.0.4
 pamqp==3.2.1
 python-decouple==3.8
 yarl==1.8.2
+pydantic
+python-dotenv==0.21.0
